consolidate extraction logic
