+++ packages/houdini/src/codegen/generators/index.ts
@@ -3,3 +3,4 @@ export { default as runtime } from './runtime'
 export { default as typescript } from './typescript'
 export { default as persistOutput } from './persistedQueries'
 export { default as definitions } from './definitions'
+export { default as indexFile } from './indexFile'
