+++ packages/houdini/src/codegen/generators/runtime/index.test.ts
@@ -5,12 +5,9 @@ import * as typeScriptParser from 'recast/parsers/typescript'
 import { test, expect } from 'vitest'
 
 import { runPipeline } from '../..'
-import { readFile, testConfig } from '../../../../lib'
-import { mockCollectedDoc } from '../../testUtils'
-import { CollectedGraphQLDocument } from '../../types'
-
-// the config to use in tests
-const config = testConfig()
+import * as fs from '../../../lib/fs'
+import { CollectedGraphQLDocument } from '../../../lib/types'
+import { testConfig, mockCollectedDoc } from '../../../test'
 
 // the documents to test
 const docs: CollectedGraphQLDocument[] = [
@@ -18,32 +15,13 @@ const docs: CollectedGraphQLDocument[] = [
 	mockCollectedDoc(`fragment TestFragment on User { firstName }`),
 ]
 
-test('index file - esm', async function () {
-	const config = testConfig({ module: 'esm' })
-
+test('runtime index file - sapper', async function () {
+	const config = testConfig({ module: 'commonjs' })
 	// execute the generator
 	await runPipeline(config, docs)
 
 	// open up the index file
-	const queryContents = await readFile(path.join(config.artifactDirectory, 'index.js'))
-	expect(queryContents).toBeTruthy()
-	// parse the contents
-	const parsedQuery: ProgramKind = recast.parse(queryContents!, {
-		parser: typeScriptParser,
-	}).program
-	// verify contents
-	expect(parsedQuery).toMatchInlineSnapshot(`
-		export { default as TestFragment} from './TestFragment'
-		export { default as TestQuery} from './TestQuery'
-	`)
-})
-
-test('index file - commonjs', async function () {
-	// execute the generator
-	await runPipeline(testConfig({ module: 'commonjs' }), docs)
-
-	// open up the index file
-	const queryContents = await readFile(path.join(config.artifactDirectory, 'index.js'))
+	const queryContents = await fs.readFile(path.join(config.rootDir, 'index.js'))
 	expect(queryContents).toBeTruthy()
 	// parse the contents
 	const parsedQuery: ProgramKind = recast.parse(queryContents!, {
@@ -66,9 +44,31 @@ test('index file - commonjs', async function () {
 		    return (mod && mod.__esModule) ? mod : { "default": mod };
 		};
 		Object.defineProperty(exports, "__esModule", { value: true });
-		var TestFragment = require("./TestFragment");
-		Object.defineProperty(exports, "TestFragment", { enumerable: true, get: function () { return __importDefault(TestFragment).default; } });
-		var TestQuery = require("./TestQuery");
-		Object.defineProperty(exports, "TestQuery", { enumerable: true, get: function () { return __importDefault(TestQuery).default; } });
+
+		__exportStar(require("./runtime"), exports);
+		__exportStar(require("./artifacts"), exports);
+		__exportStar(require("./graphql"), exports);
+		__exportStar(require("./stores"), exports);
+	`)
+})
+
+test('runtime index file - kit', async function () {
+	const config = testConfig({ module: 'esm', framework: 'kit' })
+	// execute the generator
+	await runPipeline(config, docs)
+
+	// open up the index file
+	const queryContents = await fs.readFile(path.join(config.rootDir, 'index.js'))
+	expect(queryContents).toBeTruthy()
+	// parse the contents
+	const parsedQuery: ProgramKind = recast.parse(queryContents!, {
+		parser: typeScriptParser,
+	}).program
+	// verify contents
+	expect(parsedQuery).toMatchInlineSnapshot(`
+		export * from "./runtime"
+		export * from "./artifacts"
+		export * from "./stores"
+		export * from "./graphql"
 	`)
 })
